+++ V2/CourseTypes/CourseTypesController.php
@@ -4,7 +4,7 @@ namespace IpswichJAFFARunningClubAPI\V2\CourseTypes;
 
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/BaseController.php';
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/IRoute.php';
-require_once 'CourseTypesDataAccess.php';
+require_once 'CourseTypesCommand.php';
 
 use IpswichJAFFARunningClubAPI\V2\BaseController as BaseController;
 use IpswichJAFFARunningClubAPI\V2\IRoute as IRoute;
@@ -13,21 +13,14 @@ class CourseTypesController extends BaseController implements IRoute
 {
 	public function __construct(string $route, $db)
 	{
-		parent::__construct($route, new CourseTypesDataAccess($db));
+		parent::__construct($route, new CourseTypesCommand($db));
 	}
 
 	public function registerRoutes()
 	{
 		register_rest_route($this->route, '/coursetypes', array(
 			'methods'             => \WP_REST_Server::READABLE,
-			'callback'            => array($this, 'getCourseTypes')
+			'callback'            => array($this->command, 'getCourseTypes')
 		));
-	}
-
-	public function getCourseTypes(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->getCourseTypes();
-
-		return rest_ensure_response($response);
 	}
 }
