+++ V2/Genders/GendersController.php
@@ -4,7 +4,7 @@ namespace IpswichJAFFARunningClubAPI\V2\Genders;
 
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/BaseController.php';
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/IRoute.php';
-require_once 'GendersDataAccess.php';
+require_once 'GendersCommand.php';
 
 use IpswichJAFFARunningClubAPI\V2\BaseController as BaseController;
 use IpswichJAFFARunningClubAPI\V2\IRoute as IRoute;
@@ -13,7 +13,7 @@ class GendersController extends BaseController implements IRoute
 {
 	public function __construct(string $route, $db)
 	{
-		parent::__construct($route, new GendersDataAccess($db));
+		parent::__construct($route, new GendersCommand($db));
 	}
 
 	public function registerRoutes()
@@ -21,14 +21,7 @@ class GendersController extends BaseController implements IRoute
 		register_rest_route($this->route, '/genders', array(
 			'methods' => \WP_REST_Server::READABLE,
 			'permission_callback' => array($this, 'isAuthorized'),
-			'callback' => array($this, 'getGenders')
+			'callback' => array($this->command, 'getGenders')
 		));
-	}
-
-	public function getGenders(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->getGenders();
-
-		return rest_ensure_response($response);
 	}
 }
