+++ V2/GrandPrix/GrandPrixDataAccess.php
@@ -15,26 +15,27 @@ class GrandPrixDataAccess extends DataAccess
         $sql = "SELECT
                   p.id as runnerId,
                   p.name,
-				  p.dob as dateOfBirth,
+				          p.dob as dateOfBirth,
                   ra.id as raceId,
-				  e.id as eventId,
-				  e.name as eventName,
-				  ra.description,
-				  ra.distance_id as distanceId,
-				  r.position as position,
-				  r.result as result
+                  e.id as eventId,
+                  e.name as eventName,
+                  ra.description,
+                  ra.distance_id as distanceId,
+                  r.position as position,
+                  r.result as result,
+                  r.performance as performance
                 FROM
                   results r,
                   race ra,
                   runners p,
-				  events e
+				          events e
                 WHERE
                   ra.date >= '$year-04-01' and ra.date < '$nextYear-04-01'
                   AND r.runner_id = p.id
                   AND $sexId = p.sex_id
                   AND ra.id = r.race_id
                   AND ra.grand_prix = 1
-				  AND e.id = ra.event_id
+				          AND e.id = ra.event_id
                   AND year(FROM_DAYS(TO_DAYS(ra.date) - TO_DAYS(p.dob))) >= 16
                 ORDER BY ra.date, ra.id, r.position asc, r.result asc";
 
@@ -46,13 +47,13 @@ class GrandPrixDataAccess extends DataAccess
         $raceIdString = join(",", $raceIds);
 
         $sql = "SELECT ra.id, e.id AS eventId, e.Name as eventName, ra.description as description, ra.date, ra.course_type_id AS courseTypeId, c.description AS courseType, ra.area, ra.county, ra.country_code AS countryCode, ra.conditions, ra.venue, d.distance, ra.grand_prix as isGrandPrixRace, ra.course_number as courseNumber, ra.meeting_id as meetingId, d.result_measurement_unit_type_id as resultMeasurementUnitTypeId
-			FROM `events` e
-			INNER JOIN `race` ra ON ra.event_id = e.id
-			LEFT JOIN `distance` d ON ra.distance_id = d.id
-			LEFT JOIN `course_type` c ON ra.course_type_id = c.id
-			WHERE ra.id in ($raceIdString)
-			ORDER BY ra.date";
+                FROM `events` e
+                INNER JOIN `race` ra ON ra.event_id = e.id
+                LEFT JOIN `distance` d ON ra.distance_id = d.id
+                LEFT JOIN `course_type` c ON ra.course_type_id = c.id
+                WHERE ra.id in ($raceIdString)
+                ORDER BY ra.date";
 
-		return $this->executeResultsQuery(__METHOD__, $sql);
+		    return $this->executeResultsQuery(__METHOD__, $sql);
     }
 }
