+++ V2/HistoricRecords/HistoricRecordsDataAccess.php
@@ -20,17 +20,19 @@ class HistoricRecordsDataAccess extends DataAccess
             c.id as 'categoryId',
             c.code as 'categoryCode',
             r.result,
+            r.performance,
             r.position,
             ra.course_type_id as 'courseTypeId',
-            d.result_measurement_unit_type_id as resultMeasurementUnitTypeId
+            d.result_measurement_unit_type_id as resultMeasurementUnitTypeId,
+            d.result_unit_type_id as resultUnitTypeId
             FROM `results` r
             inner join race ra on ra.id = r.race_id
             INNER JOIN runners p ON p.id = r.runner_id
             INNER JOIN events e ON e.id = ra.event_id
             INNER JOIN category c ON c.id = r.category_id
             INNER JOIN distance d ON d.id = ra.distance_id
-            WHERE ra.distance_id = %d AND c.id > 0 AND r.result <> '00:00:00' AND r.result <> ''
-            order by category_id asc, ra.date asc, r.result asc",
+            WHERE ra.distance_id = %d AND c.id > 0 AND r.performance IS NOT NULL
+            order by category_id asc, ra.date asc, r.performance asc",
             $distanceId
         );
 
@@ -49,19 +51,21 @@ class HistoricRecordsDataAccess extends DataAccess
             c.id as 'categoryId',
             c.code as 'categoryCode',
             r.result,
+            r.performance,
             r.position,
             ra.course_type_id as 'courseTypeId',
             d.distance,
             d.id as 'distanceId',
-            d.result_measurement_unit_type_id as resultMeasurementUnitTypeId
+            d.result_measurement_unit_type_id as resultMeasurementUnitTypeId,
+            d.result_unit_type_id as resultUnitTypeId
             FROM `results` r
             INNER join race ra on ra.id = r.race_id
             INNER JOIN runners p ON p.id = r.runner_id
             INNER JOIN events e ON e.id = ra.event_id
             INNER JOIN category c ON c.id = r.category_id
             INNER JOIN distance d ON d.id = ra.distance_id
-            WHERE r.category_id = %d AND r.result <> '00:00:00' AND r.result <> ''
-            order by d.id asc, ra.date asc, r.result asc",
+            WHERE r.category_id = %d AND r.performance IS NOT NULL
+            order by d.id asc, ra.date asc, r.performance asc",
             $categoryId
         );
