+++ V2/RunnerResults/RunnerResultsDataAccess.php
@@ -75,8 +75,8 @@ class RunnerResultsDataAccess extends DataAccess
 					  ra.description as raceName,
 					  ra.id as raceId,
 					  r.position as position,
-					  r.result as time,
 					  r.result as result,
+					  r.performance as performance,
 					  r.personal_best as isPersonalBest,
 					  r.season_best as isSeasonBest,
 					  st.name as standard,
@@ -117,8 +117,8 @@ class RunnerResultsDataAccess extends DataAccess
 					  ra.description as raceName,
 					  ra.id as raceId,
 					  r.position as position,
-					  r.result as time,
 					  r.result as result,
+					  r.performance as performance,
 					  r.info as info,
 					  CASE
 					   WHEN ra.date >= '%s' THEN r.percentage_grading_2015
@@ -142,8 +142,7 @@ class RunnerResultsDataAccess extends DataAccess
 						where
 						  r.runner_id = %d
 						  and r.personal_best = 1
-						  and r.result != '00:00:00'
-						  and r.result != ''
+						  and r.performance IS NOT NULL
 						group by ra.distance_id
 					) t on r.result = t.pb and ra.distance_id = t.distanceId
 					where
