+++ V2/Runners/RunnersController.php
@@ -4,30 +4,28 @@ namespace IpswichJAFFARunningClubAPI\V2\Runners;
 
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/BaseController.php';
 require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/IRoute.php';
-require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/Distances/Distances.php';
-require_once 'RunnersDataAccess.php';
+require_once 'RunnersCommand.php';
 
 use IpswichJAFFARunningClubAPI\V2\BaseController as BaseController;
 use IpswichJAFFARunningClubAPI\V2\IRoute as IRoute;
-use IpswichJAFFARunningClubAPI\V2\Distances\Distances as Distances;
 
 class RunnersController extends BaseController implements IRoute
 {
 	public function __construct(string $route, $db)
 	{
-		parent::__construct($route, new RunnersDataAccess($db));
+		parent::__construct($route, new RunnersCommand($db));
 	}
 
 	public function registerRoutes()
 	{
 		register_rest_route($this->route, '/runners', array(
 			'methods' => \WP_REST_Server::READABLE,
-			'callback' => array($this, 'getRunners')
+			'callback' => array($this->command, 'getRunners')
 		));
 
 		register_rest_route($this->route, '/runners/(?P<runnerId>[\d]+)', array(
 			'methods' => \WP_REST_Server::READABLE,
-			'callback' => array($this, 'getRunner'),
+			'callback' => array($this->command, 'getRunner'),
 			'args' => array(
 				'runnerId' => array(
 					'required' => true,
@@ -39,7 +37,7 @@ class RunnersController extends BaseController implements IRoute
 		register_rest_route($this->route, '/runners', array(
 			'methods' => \WP_REST_Server::CREATABLE,
 			'permission_callback' => array($this, 'isAuthorized'),
-			'callback' => array($this, 'saveRunner'),
+			'callback' => array($this->command, 'saveRunner'),
 			'args' => array(
 				'runner' => array(
 					'required' => true,
@@ -51,7 +49,7 @@ class RunnersController extends BaseController implements IRoute
 		register_rest_route($this->route, '/runners/(?P<runnerId>[\d]+)', array(
 			'methods' => \WP_REST_Server::DELETABLE,
 			'permission_callback' => array($this, 'isAuthorized'),
-			'callback' => array($this, 'deleteRunner'),
+			'callback' => array($this->command, 'deleteRunner'),
 			'args' => array(
 				'runnerId' => array(
 					'required' => true,
@@ -63,7 +61,7 @@ class RunnersController extends BaseController implements IRoute
 		register_rest_route($this->route, '/runners/(?P<runnerId>[\d]+)', array(
 			'methods' => \WP_REST_Server::EDITABLE,
 			'permission_callback' => array($this, 'isAuthorized'),
-			'callback' => array($this, 'updateRunner'),
+			'callback' => array($this->command, 'updateRunner'),
 			'args' => array(
 				'runnerId' => array(
 					'required' => true,
@@ -78,58 +76,7 @@ class RunnersController extends BaseController implements IRoute
 				)
 			)
 		));
-	}
-
-	public function getRunners(\WP_REST_Request $request)
-	{
-		$loggedIn = $this->isAuthorized($request);
-		$response = $this->dataAccess->getRunners($loggedIn);
-
-		return rest_ensure_response($response);
-	}
-
-	public function getRunner(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->getRunner($request['runnerId']);
-		$certificates = $this->dataAccess->getStandardCertificates($request['runnerId']);
-		$distances = array(
-			Distances::FIVE_KILOMETRES,
-			Distances::FIVE_MILES,
-			Distances::TEN_KILOMETRES,
-			Distances::TEN_MILES,
-			Distances::HALF_MARATHON,
-			Distances::TWENTY_MILES,
-			Distances::MARATHON
-		);
-
-		$rankings = $this->dataAccess->getRunnerRankings($request['runnerId'], $response->sexId, $distances);
-
-		$response->certificates = $certificates;
-		$response->rankings = $rankings;
-
-		return rest_ensure_response($response);
-	}
-
-	public function saveRunner(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->insertRunner($request['runner']);
-
-		return rest_ensure_response($response);
-	}
-
-	public function deleteRunner(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->deleteRunner($request['runnerId']);
-
-		return rest_ensure_response($response);
-	}
-
-	public function updateRunner(\WP_REST_Request $request)
-	{
-		$response = $this->dataAccess->updateRunner($request['runnerId'], $request['field'], $request['value']);
-
-		return rest_ensure_response($response);
-	}
+	}	
 
 	public function isValidRunnerUpdateField($value, $request, $key)
 	{
