+++ V2/TeamResults/TeamResultsDataAccess.php
@@ -23,7 +23,7 @@ class TeamResultsDataAccess extends DataAccess
     public function getTeamResult(int $teamResultId)
     {
         $sql = $this->resultsDatabase->prepare(
-            'SELECT tr.id as teamId, p.name as runnerName, p.id as runnerId, r.performance as runnerResult,
+            'SELECT tr.id as teamId, p.name as runnerName, p.id as runnerId, r.result as runnerResult,
 			r.position as runnerPosition, trr.order as teamOrder
 			FROM `team_results` tr
 			INNER JOIN `team_results_runners` trr ON tr.id = trr.team_result_id
@@ -40,10 +40,10 @@ class TeamResultsDataAccess extends DataAccess
     public function insertTeamResult(int $teamResult)
     {
         $sql = $this->resultsDatabase->prepare(
-            'INSERT INTO team_results (`team_name`, `category`, `performance`, `position`, `meeting_id`) VALUES(%s, %s, %s, %d, %d);',
+            'INSERT INTO team_results (`team_name`, `category`, `result`, `position`, `meeting_id`) VALUES(%s, %s, %s, %d, %d);',
             $teamResult['name'],
             $teamResult['category'],
-            $teamResult['performance'],
+            $teamResult['result'],
             $teamResult['position'],
             $teamResult['meetingId']
         );
